+++ bee-ledger/src/workers/pruning/constants.rs
@@ -1,6 +1,0 @@
-// Copyright 2020-2021 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-pub(crate) const SOLID_ENTRY_POINT_THRESHOLD_PAST: u32 = 5;
-pub(crate) const SOLID_ENTRY_POINT_THRESHOLD_FUTURE: u32 = 5;
-pub(crate) const PRUNING_THRESHOLD: u32 = 50;