+++ bee-ledger/src/workers/snapshot/mod.rs
@@ -3,6 +3,7 @@
 
 //! Module containing snapshot operations.
 
+pub(crate) mod condition;
 pub(crate) mod download;
 pub(crate) mod error;
 pub(crate) mod import;