+++ bee-storage/bee-storage-rocksdb/tests/milestone_index_to_output_diff.rs
@@ -4,4 +4,4 @@
 #[macro_use]
 mod access;
 
-impl_access_test!(milestone_index_to_output_diff_access);
+// impl_access_test!(milestone_index_to_output_diff_access);