+++ bee-storage/bee-storage-test/src/lib.rs
@@ -1,7 +1,7 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-mod address_to_balance;
+// mod address_to_balance;
 mod ed25519_address_to_output_id;
 mod index_to_message_id;
 mod ledger_index;
@@ -9,7 +9,7 @@ mod message_id_to_message;
 mod message_id_to_message_id;
 mod message_id_to_metadata;
 mod milestone_index_to_milestone;
-mod milestone_index_to_output_diff;
+// mod milestone_index_to_output_diff;
 mod milestone_index_to_receipt;
 mod milestone_index_to_unreferenced_message;
 mod output_id_to_consumed_output;
@@ -19,7 +19,7 @@ mod snapshot_info;
 mod solid_entry_point_to_milestone_index;
 mod spent_to_treasury_output;
 
-pub use address_to_balance::address_to_balance_access;
+// pub use address_to_balance::address_to_balance_access;
 pub use ed25519_address_to_output_id::ed25519_address_to_output_id_access;
 pub use index_to_message_id::index_to_message_id_access;
 pub use ledger_index::ledger_index_access;
@@ -27,7 +27,7 @@ pub use message_id_to_message::message_id_to_message_access;
 pub use message_id_to_message_id::message_id_to_message_id_access;
 pub use message_id_to_metadata::message_id_to_metadata_access;
 pub use milestone_index_to_milestone::milestone_index_to_milestone_access;
-pub use milestone_index_to_output_diff::milestone_index_to_output_diff_access;
+// pub use milestone_index_to_output_diff::milestone_index_to_output_diff_access;
 pub use milestone_index_to_receipt::milestone_index_to_receipt_access;
 pub use milestone_index_to_unreferenced_message::milestone_index_to_unreferenced_message_access;
 pub use output_id_to_consumed_output::output_id_to_consumed_output_access;